RChain - added rChain service/manager/repository
