+++ server/modules/analysis/repository/processingChainRepository.js
@@ -70,31 +70,6 @@ export const fetchChainByUuid = async (surveyId, processingChainUuid, client = d
     dbTransformCallback,
   )
 
-export const fetchStepData = async (survey, cycle, step, client = db) => {
-  const entityDef = Survey.getNodeDefByUuid(ProcessingStep.getEntityUuid(step))(survey)
-  const entityDefParent = Survey.getNodeDefParent(entityDef)(survey)
-  const viewName = NodeDefTable.getViewName(entityDef, entityDefParent)
-  const calculationAttrDefs = R.pipe(
-    ProcessingStep.getCalculations,
-    R.map(
-      R.pipe(ProcessingStepCalculation.getNodeDefUuid, nodeDefUuid => Survey.getNodeDefByUuid(nodeDefUuid)(survey)),
-    ),
-  )(step)
-
-  const fields = ['*']
-  for (const nodeDef of calculationAttrDefs) {
-    const nodeDefName = NodeDef.getName(nodeDef)
-    // Add nodeDefName_uuid field
-    fields.push(`uuid_generate_v4() as ${nodeDefName}_uuid`)
-  }
-
-  return await client.any(`
-      SELECT ${fields.join(', ')} 
-      FROM ${SchemaRdb.getName(Survey.getId(survey))}.${viewName}
-      WHERE ${DataTable.colNameRecordCycle} = '${cycle}' 
-    `)
-}
-
 // ====== UPDATE
 
 export const updateChainProp = async (surveyId, processingChainUuid, key, value, client = db) =>
