+++ server/modules/analysis/service/processingChainService.js
@@ -1,8 +1,5 @@
 import RChain from '@server/modules/analysis/service/_rChain/rChain'
 
-import * as SurveyManager from '@server/modules/survey/manager/surveyManager'
-import * as ProcessingChainManager from '@server/modules/analysis/manager/processingChainManager'
-
 export {
   // ======  READ - Chain
   countChainsBySurveyId,
@@ -25,11 +22,6 @@ export {
   // ======  DELETE - Calculation
   deleteCalculation,
 } from '../manager/processingChainManager'
-
-export const fetchStepData = async (surveyId, cycle, stepUuid) => {
-  const survey = await SurveyManager.fetchSurveyAndNodeDefsBySurveyId(surveyId, cycle)
-  return await ProcessingChainManager.fetchStepData(survey, cycle, stepUuid)
-}
 
 export const generateScript = async (surveyId, cycle, chainUuid, serverUrl) => {
   await new RChain(surveyId, cycle, chainUuid, serverUrl).init()
