+++ server/system/apiRouter.js
@@ -11,6 +11,7 @@ import * as collectImportApi from '@server/modules/collectImport/api/collectImpo
 import * as recordApi from '@server/modules/record/api/recordApi'
 import * as surveyRdbApi from '@server/modules/surveyRdb/api/surveyRdbApi'
 import * as processingChainApi from '@server/modules/analysis/api/processingChainApi'
+import * as rChainApi from '@server/modules/analysis/api/rChainApi'
 import * as jobApi from '@server/job/jobApi'
 import * as expressionApi from '@server/modules/expression/api/expressionApi'
 
@@ -34,6 +35,8 @@ recordApi.init(router)
 surveyRdbApi.init(router)
 
 processingChainApi.init(router)
+rChainApi.init(router)
+
 jobApi.init(router)
 
 expressionApi.init(router)
